migrate create-studiocms to new clikit
