+++ packages/create-studiocms/package.json
@@ -26,7 +26,7 @@
 		"@bluwy/giget-core": "^0.1.2",
 		"@clack/core": "0.4.1",
 		"@clack/prompts": "0.9.1",
-		"@commander-js/extra-typings": "^13.0.0",
+		"@withstudiocms/cli-kit": "workspace:*",
 		"commander": "^13.0.0",
 		"figlet": "^1.8.0",
 		"chalk": "^5.4.1",
