+++ packages/create-studiocms/src/cmds/get-turso.ts
@@ -1,10 +1,12 @@
-import { Command } from '../utils/commander.js';
-import { logger, runInteractiveCommand } from '../utils/index.js';
+import { Command } from '@withstudiocms/cli-kit/commander';
+import { buildLogger } from '@withstudiocms/cli-kit/logger';
+import { runInteractiveCommand } from '@withstudiocms/cli-kit/utils';
 
 await new Command('getTurso')
 	.description('Turso CLI Installer')
 	.summary('Turso CLI Installer')
 	.action(async () => {
+		const logger = buildLogger(false);
 		try {
 			logger.log('Starting Turso install...');
 			await runInteractiveCommand('curl -sSfL https://get.tur.so/install.sh | bash');
