+++ packages/create-studiocms/src/cmds/interactive/template.ts
@@ -1,8 +1,8 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { downloadTemplate } from '@bluwy/giget-core';
+import { StudioCMSColorwayError, StudioCMSColorwayInfo } from '@withstudiocms/cli-kit/colors';
 import color from 'chalk';
-import { StudioCMSColorwayError, StudioCMSColorwayInfo } from '../../utils/index.js';
 import type { Context } from './context.js';
 
 function templateTargetFilter(
